requirements.txt
rank_bm25
ollama
trafilatura
sentence-transformers
spacy
pyyaml
requests
trafilatura
python-dotenv